Changed storage method for ease of access

+++ frontend/toolkit-webapp/src/pages/ModuleCreator.js
@@ -29,34 +29,12 @@ function ModuleCreator(props) {
             return
         }
 
-        if (type === 'Quiz') {
+        if (type === 'Quiz' && JSON.parse(sessionStorage.getItem('quiz')) !== null) {
             console.log('works for Quiz')
-            var quizArray = {
-                questions: JSON.parse(sessionStorage.getItem('questions')),
-                types: JSON.parse(sessionStorage.getItem('types')),
-                answers: JSON.parse(sessionStorage.getItem('answers')),
-                fakes1: JSON.parse(sessionStorage.getItem('fakes1')),
-                fakes2: JSON.parse(sessionStorage.getItem('fakes2')),
-                fakes3: JSON.parse(sessionStorage.getItem('fakes3'))
-            }
             var quiz =[]
 
-            for(var i = 0; i < quizArray.questions.length; i++){
-                quiz.push({
-                    question: quizArray.questions[i],
-                    type: quizArray.types[i],
-                    answers: [
-                        {answerText: quizArray.answers[i], isCorrect: true},
-                        {answerText: quizArray.fakes1[i], isCorrect: false},
-                        {answerText: quizArray.fakes2[i], isCorrect: false},
-                        {answerText: quizArray.fakes3[i], isCorrect: false},
-                    ]
-
-                })
-            }
-            
+            quiz = JSON.parse(sessionStorage.getItem("quiz"))
             sessionStorage.clear()
-            //console.log(quiz)
             onFinish({ title, type, description, quiz })
         } else {
             console.log('works')
@@ -80,7 +58,6 @@ function ModuleCreator(props) {
             let res = undefined
             if (module.type === "Quiz") {
                 res = await fetch(config.server_url + config.paths.newModule, {
-
                     method: 'POST',
                     headers: {
                         'Content-type': 'application/json'
@@ -112,11 +89,6 @@ function ModuleCreator(props) {
         else {
             props.history.push('login')
         }
-
-    }
-
-    const getFileModule = () => {
-        return <FileModule></FileModule>
 
     }
 
@@ -185,10 +157,10 @@ function ModuleCreator(props) {
                             <Container className={classes.buttonGroup}>
                                 <Button type='submit' className={classes.button1} size="small" variant="contained" onClick={cancel}>
                                     Cancel
-                        </Button>
+                                </Button>
                                 <Button type='submit' className={classes.button2} size="small" variant="contained" onSubmit={onSubmit}>
                                     Create
-                        </Button>
+                                </Button>
                             </Container>
                         </Paper>
                     </form>
